Product

+++ src/main/java/com/bmlinformtica/course/entities/Order.java
@@ -23,7 +23,7 @@ public class Order implements Serializable {
 	// Propriedades da classe order
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 
 	@JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss'Z'", timezone = "GMT")
 	private Instant moment;
@@ -40,7 +40,7 @@ public class Order implements Serializable {
 	}
 
 	// Método de instanciação das propriedades da classe
-	public Order(Integer id, Instant moment, OrderStatus orderStatus, User client) {
+	public Order(Long id, Instant moment, OrderStatus orderStatus, User client) {
 		super();
 		this.id = id;
 		this.moment = moment;
@@ -49,11 +49,11 @@ public class Order implements Serializable {
 	}
 
 	// Getters and Setters
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
