+++ src/main/java/com/bmlinformtica/course/repositories/OrderRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.bmlinformtica.course.entities.Order;
 
-public interface OrderRepository extends JpaRepository<Order, Integer>{
+public interface OrderRepository extends JpaRepository<Order, Long>{
 
 }
